fix: fix modal duplication, add .me()
